finalized project with TCKN update

+++ Efectura.Data/Model/Consumer.cs
@@ -6,7 +6,7 @@ namespace Efectura.Data.Model
 {
     public class Consumer : BaseEntity
     {
-        public int TCKN { get; set; }
+        public double TCKN { get; set; }
         public string Name { get; set; }
         public string SurName { get; set; }
         public DateTime BirthDate { get; set; }
